feat(GlobeController): terrain-aware default via withTerrain mixin
Extract terrain-following into a reusable withTerrain(Base) mixin so GlobeController
is terrain-aware WITHOUT inheriting MapController (removing the maxBounds/normalize
band-aid). TerrainController becomes withTerrain(MapController); public API unchanged.

Move rotationPivot + _getAltitude to the base Controller so the globe inherits
rotate-around-pivot; MapController keeps only its mercator position/maxBounds in setProps.
Add GlobeViewport.panByPosition3D for 3D-anchored panning.
